Shared memory, textbox has blinking cursor and minor fixes

+++ Kernel/include/string.h
@@ -19,7 +19,7 @@ int strcmp(const char* s1, const char* s2);
 
 char *strtok(char * str, const char * delim);
 
-int strlen(const char* str);
+size_t strlen(const char* str);
 char* strcat(char* dest, const char* src);
 char* strncat(char* dest, const char* src, size_t n);
